chore(notebooks): migrate models from posthog app to products structure
Moves notebooks models (Notebook, ResourceNotebook) from posthog/models/notebook/
to products/notebooks/backend/ to align with the product-oriented architecture.

Uses SeparateDatabaseAndState pattern to update Django's internal state without
modifying database tables. Updates ContentType records to point to new app label
while keeping db_table references to original table names.

Updates all imports across codebase (23 files) to use new import path.
Moves notebook API endpoints from posthog/api/ to products/notebooks/backend/api/.

Related: model migration workflow, notebooks product structure
