+++ posthog/api/search.py
@@ -12,7 +12,8 @@ from rest_framework.response import Response
 from posthog.api.routing import TeamAndOrgViewSetMixin
 from posthog.helpers.full_text_search import build_rank, process_query
 from posthog.models import Action, Cohort, Dashboard, EventDefinition, Experiment, FeatureFlag, Insight, Survey
-from posthog.models.notebook.notebook import Notebook
+
+from products.notebooks.backend.models import Notebook
 
 LIMIT = 25
