+++ posthog/api/test/notebooks/test_notebook_filtering.py
@@ -6,7 +6,8 @@ from parameterized import parameterized
 from rest_framework import status
 
 from posthog.models import User
-from posthog.models.notebook.notebook import Notebook
+
+from products.notebooks.backend.models import Notebook
 
 PLAYLIST_CONTENT = lambda: {
     "type": "ph-recording-playlist",
