+++ posthog/api/test/test_search.py
@@ -4,8 +4,10 @@ from posthog.test.base import APIBaseTest
 from django.db import connection
 
 from posthog.helpers.full_text_search import process_query
-from posthog.models import Dashboard, FeatureFlag, Insight, Notebook, Team
+from posthog.models import Dashboard, FeatureFlag, Insight, Team
 from posthog.models.event_definition import EventDefinition
+
+from products.notebooks.backend.models import Notebook
 
 
 class TestSearch(APIBaseTest):
