+++ posthog/models/file_system/unfiled_file_saver.py
@@ -14,13 +14,13 @@ from posthog.models.file_system.file_system_mixin import FileSystemSyncMixin
 from posthog.models.hog_functions.hog_function import HogFunction
 from posthog.models.insight import Insight
 from posthog.models.link import Link
-from posthog.models.notebook import Notebook
 from posthog.models.surveys.survey import Survey
 from posthog.models.team import Team
 from posthog.models.user import User
 from posthog.session_recordings.models.session_recording_playlist import SessionRecordingPlaylist
 
 from products.early_access_features.backend.models import EarlyAccessFeature
+from products.notebooks.backend.models import Notebook
 
 MIXIN_MODELS = {
     "action": Action,
