--- posthog/models/notebook/__init__.py
@@ -1,2 +1,0 @@
-from .notebook import *
-from .resource_notebook import *
