+++ products/notebooks/backend/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class NotebooksConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "products.notebooks.backend"
+    label = "notebooks"
