+++ tach.toml
@@ -36,7 +36,8 @@ depends_on = [
     "products.marketing_analytics",
     "products.endpoints",
 
-    "products.error_tracking",]
+    "products.error_tracking",
+    "products.notebooks",]
 
 [[modules]]
 path = "products.early_access_features"
@@ -64,4 +65,8 @@ depends_on = ["posthog"]
 
 [[modules]]
 path = "products.error_tracking"
+depends_on = ["posthog"]
+
+[[modules]]
+path = "products.notebooks"
 depends_on = ["posthog"]
